+++ kvdb-memorydb/CHANGELOG.md
@@ -11,7 +11,7 @@ The format is based on [Keep a Changelog].
 
 ## [0.2.0] - 2019-12-19
 ### Fixed
-- `iter_from_prefix` behaviour synced with the `kvdb-rocksdb`
+- `iter_from_prefix` behaviour synced with the `tetsy-kvdb-rocksdb`
 
 ### Changed
 - Default column support removed from the API